Second block (does not compile yet)

+++ src/lib/operators/table_scan/abstract_table_scan_impl.hpp
@@ -139,7 +139,7 @@ class AbstractTableScanImpl {
         const auto& left = *left_it;
 
         if constexpr (std::is_same_v<RightIterator, std::false_type>) {
-          mask |= ((!CheckForNull | !left.is_null()) && func(left)) << index;
+          mask |= static_cast<uint64_t>((!CheckForNull | !left.is_null()) && func(left)) << index;
         } else {
           const auto& right = *right_it;
           mask |= ((!CheckForNull | (!left.is_null() && !right.is_null())) & func(left, right)) << index;
